move data_module wraper to another module
