+++ harness/determined/experimental/pytorch_lightning/__init__.py
@@ -1,6 +1,8 @@
 from determined.experimental.pytorch_lightning.adapter import (
-    DETLightningDataModule,
     DETLightningModule,
     HyperparamsProvider,
     PTLAdapter,
+)
+from determined.experimental.pytorch_lightning.data_module import (
+    DETLightningDataModule,
 )
